% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalMuFunctions.R
\name{FS}
\alias{FS}
\title{Final Size Probability}
\usage{
FS(y, R0, mu)
}
\arguments{
\item{y}{Final size, a positive integer}

\item{R0}{Reproduction number, a positive number}

\item{mu}{Mutation rate, a positive number}
}
\value{
A number representing the probability
}
\description{
This function calculates the probability of getting a specific final size y
with initial population size of 1 and some certain parameters.
}
\examples{
FS(3, 1.2, 0.4)
FS(6, 3.5, 1.6)
}
