% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalMuFunctions.R
\name{FSn}
\alias{FSn}
\title{Final Size Probability with Arbitrary Initial Population Size}
\usage{
FSn(n, y, R0, mu)
}
\arguments{
\item{n}{Initial population size, a positive integer}

\item{y}{Final size, a positive integer}

\item{R0}{Reproduction number, a positive number}

\item{mu}{Mutation rate, a positive number}
}
\value{
A number representing the probability
}
\description{
This function calculates the probability of getting a specific final size y
with initial population size of N and some certain parameters.
}
\examples{
FSn(2, 3, 1.2, 0.4)
FSn(3, 6, 3.5, 1.6)
}
