% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalMuFunctions.R
\name{list_clonal}
\alias{list_clonal}
\title{Clonal Mutations with Various Parameters}
\usage{
list_clonal(n_values, R0, mu_values, maxMuGen, maxFS, clonal)
}
\arguments{
\item{n_values}{A sequence of initial population sizes}

\item{R0}{Reproduction number, a positive number}

\item{mu_values}{A sequence of mutation rate}

\item{maxMuGen}{Maximum number of mutant lineages being calculated}

\item{maxFS}{Maximum number of final sizes being calculated}

\item{clonal}{Maximum number of clonal mutations being calculated}
}
\value{
A list of clonal mutation pmf with different parameters
}
\description{
This function calculates for the clonal mutation distribution
across pairs of parameters N and mu.
}
\examples{
list_clonal(1:10, 1.6, seq(0.1, 2, by = 0.2), 50, 50, 5)
list_clonal(seq(2, 10, by = 2), 0.8, seq(1.2, 2.5, by = 0.1), 60, 120, 7)
}
