% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalMuFunctions.R
\name{LL_lambda}
\alias{LL_lambda}
\title{Lambda Log Likelihood}
\usage{
LL_lambda(df, listClonal, lambda, R0, mu, maxMuGen, maxFS, maxIni)
}
\arguments{
\item{df}{Data set with first column named "ClonalMu" recording number of clonal mutations, and the second column named "freq" recording number of times the clonal mutation is observed}

\item{listClonal}{Returned list from function list_clonal with same parameters}

\item{lambda}{Mean of Poisson distribution, a positive number}

\item{R0}{Reproduction number, a positive number}

\item{mu}{Mutation rate, a positive number}

\item{maxMuGen}{Maximum number of mutant lineages being calculated}

\item{maxFS}{Maximum number of final sizes being calculated}

\item{maxIni}{Maximum initial population size being calculated}
}
\value{
A number representing the probability
}
\description{
This function calculates the log likelihood of getting the observed data
points in the provided data set given a certain lambda, R0, and mu. We
usually assume that R0 is a known value and aims to estimate lambda and mu.
}
\examples{
LL_lambda(sample_data, listClonal, 2.1, 1.6, 0.4, 50, 50, 10)
LL_lambda(flu_data, listClonal, 0.2, 10, 0.7, 60, 70, 8)
}
