% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalMuFunctions.R
\name{meanClonal.df}
\alias{meanClonal.df}
\title{Expected Clonal Mutations Across Parameters}
\usage{
meanClonal.df(n, R0_values, mu_values, maxMuGen, maxFS, maxClonal)
}
\arguments{
\item{n}{Initial population size, a positive integer}

\item{R0_values}{A sequence of reproduction number}

\item{mu_values}{A sequence of mutation rate}

\item{maxMuGen}{Maximum number of mutant lineages being calculated}

\item{maxFS}{Maximum number of final sizes being calculated}

\item{maxClonal}{Maximum number of clonal mutations being calculated}
}
\value{
A number representing the expected clonal mutations
}
\description{
This function calculates the expected number of clonal mutations for a
sequence of R0 and mu values.
}
\examples{
meanClonal.df(2, seq(1.1, 7.1, by = 0.1), seq(0.001, 0.101, by = 0.002), 60, 60, 5)
meanClonal.df(3, seq(1.05, 3, by = 0.05), seq(0.3, 0.5, by = 0.1), 120, 130, 5)
}
