% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalMuFunctions.R
\name{meanClonal}
\alias{meanClonal}
\title{Expected Clonal Mutations}
\usage{
meanClonal(n, R0, mu, maxMuGen, maxFS, maxClonal)
}
\arguments{
\item{n}{Initial population size, a positive integer}

\item{R0}{Reproduction number, a positive number}

\item{mu}{Mutation rate, a positive number}

\item{maxMuGen}{Maximum number of mutant lineages being calculated}

\item{maxFS}{Maximum number of final sizes being calculated}

\item{maxClonal}{Maximum number of clonal mutations being calculated}
}
\value{
A number representing the expected clonal mutations
}
\description{
This function calculates the expected number of clonal mutations.
}
\examples{
meanClonal(2, 1.2, 0.4, 60, 60, 5)
meanClonal(3, 3.5, 1.6, 120, 130, 5)
}
