% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalMuFunctions.R
\name{P0}
\alias{P0}
\title{Probability of Getting 0 Clonal Mutation}
\usage{
P0(n, R0, mu, maxMuGen, maxFS)
}
\arguments{
\item{n}{Initial population size, a positive integer}

\item{R0}{Reproduction number, a positive number}

\item{mu}{Mutation rate, a positive number}

\item{maxMuGen}{Maximum number of mutant lineages being calculated}

\item{maxFS}{Maximum number of final sizes being calculated}
}
\value{
A number representing the probability
}
\description{
This function calculates the probability of getting 0 clonal mutation.
}
\examples{
P0(2, 1.2, 0.4, 60, 60)
P0(3, 3.5, 1.6, 120, 130)
}
