% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalMuFunctions.R
\name{P1plus}
\alias{P1plus}
\title{Probability of Getting More Than 1 Clonal Mutation}
\usage{
P1plus(n, R0, mu, maxMuGen, maxFS)
}
\arguments{
\item{n}{Initial population size, a positive integer}

\item{R0}{Reproduction number, a positive number}

\item{mu}{Mutation rate, a positive number}

\item{maxMuGen}{Maximum number of mutant lineages being calculated}

\item{maxFS}{Maximum number of final sizes being calculated}
}
\value{
A number representing the probability
}
\description{
This function calculates the probability of getting more than 1 clonal
mutations.
}
\examples{
P1plus(2, 1.2, 0.4, 60, 60)
P1plus(3, 3.5, 1.6, 120, 130)
}
