% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalMuFunctions.R
\name{P_clonal_round}
\alias{P_clonal_round}
\title{Round Probability of Getting Clonal Mutation}
\usage{
P_clonal_round(n, m, R0, mu, maxMuGen, maxFS)
}
\arguments{
\item{n}{Initial population size, a positive integer}

\item{m}{Number of clonal mutation, a positive integer}

\item{R0}{Reproduction number, a positive number}

\item{mu}{Mutation rate, a positive number}

\item{maxMuGen}{Maximum number of mutant lineages being calculated}

\item{maxFS}{Maximum number of final sizes being calculated}
}
\value{
A number representing the probability
}
\description{
This function calculates the round probability of getting a certain number
of clonal mutations. Notice that this is not the exact probability.
}
\examples{
pmf.P_clonal_round(2, 3, 1.2, 0.4, 60, 60)
pmf.P_clonal_round(3, 6, 3.5, 1.6, 120, 130)
}
