% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalMuFunctions.R
\name{P_mu_est}
\alias{P_mu_est}
\title{Mutant Lineages Established Probability}
\usage{
P_mu_est(n, m, R0, mu, maxMuGen, maxFS)
}
\arguments{
\item{n}{Initial population size, a positive integer}

\item{m}{Number of mutant lineages established, a positive integer}

\item{R0}{Reproduction number, a positive number}

\item{mu}{Mutation rate, a positive number}

\item{maxMuGen}{Maximum number of mutant lineages being calculated}

\item{maxFS}{Maximum number of final sizes being calculated}
}
\value{
A number representing the probability
}
\description{
This function calculates the probability of a specific number of mutant
lineages established.
}
\examples{
P_mu_est(2, 1, 1.2, 0.4, 60, 60)
P_mu_est(3, 4, 3.5, 1.6, 120, 130)
}
