% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalMuFunctions.R
\name{P_mu_n}
\alias{P_mu_n}
\title{Mutant Lineages Generated Probability}
\usage{
P_mu_n(n, m, R0, mu, maxFS)
}
\arguments{
\item{n}{Initial population size, a positive integer}

\item{m}{Number of mutant lineages generated, a positive integer}

\item{R0}{Reproduction number, a positive number}

\item{mu}{Mutation rate, a positive number}

\item{maxFS}{Maximum number of final sizes being calculated}
}
\value{
A number representing the probability
}
\description{
This function calculates the probability of generating a specific number of
mutant lineages.
}
\examples{
P_mu_n(2, 1, 1.2, 0.4, 60)
P_mu_n(3, 4, 3.5, 1.6, 120)
}
