% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalMuFunctions.R
\name{pmf.alloff}
\alias{pmf.alloff}
\title{Offspring Distribution}
\usage{
pmf.alloff(R0, mu, maxOff)
}
\arguments{
\item{R0}{Reproduction number, a positive number}

\item{mu}{Mutation rate, a positive number}

\item{maxOff}{Maximum offspring number being calculated}
}
\value{
A data frame of offspring size, type, and probabilities
}
\description{
This function generates distribution of wild-type, mutant, and overall
offspring with a given R0, mutation rate, and a maximum number of offspring
you want to calculate for.
}
\examples{
pmf.alloff(1.2, 0.4, 5)
pmf.alloff(3.5, 1.6, 8)
}
