% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalMuFunctions.R
\name{pmf.clonal_params}
\alias{pmf.clonal_params}
\title{Clonal Mutation Probability with Various Parameters}
\usage{
pmf.clonal_params(clonal, listClonal, n_values, R0, mu_values, maxMuGen, maxFS)
}
\arguments{
\item{clonal}{The number of clonal mutations being calculated, a positive integer}

\item{listClonal}{Returned list from function list_clonal with same parameters}

\item{n_values}{A sequence of initial population sizes}

\item{R0}{Reproduction number, a positive number}

\item{mu_values}{A sequence of mutation rate}

\item{maxMuGen}{Maximum number of mutant lineages being calculated}

\item{maxFS}{Maximum number of final sizes being calculated}
}
\value{
A data frame of N, mu values and corresponding probability of getting the input clonal mutations
}
\description{
This function calculates the probability of getting a certain
number of clonal mutations across pairs of parameters N and mu.
}
\examples{
pmf.clonal_params(0, listClonal, 1:10, 1.6, seq(0.1, 2, by = 0.2), 50, 50)
pmf.clonal_params(2, listClonal, seq(2, 10, by = 2), 0.8, seq(1.2, 2.5, by = 0.1), 60, 120)
}
