% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalMuFunctions.R
\name{pmf.FS}
\alias{pmf.FS}
\title{Final Size Distribution}
\usage{
pmf.FS(R0, mu, maxFS)
}
\arguments{
\item{R0}{Reproduction number, a positive number}

\item{mu}{Mutation rate, a positive number}

\item{maxFS}{Maximum number of final sizes being calculated}
}
\value{
A data frame of final sizes and corresponding probabilities
}
\description{
This function calculates the final size distribution with initial population
size of 1 and some certain parameters.
}
\examples{
pmf.FS(1.2, 0.4, 60)
pmf.FS(3.5, 1.6, 120)
}
