% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalMuFunctions.R
\name{pmf.P_mu_est}
\alias{pmf.P_mu_est}
\title{Mutant Lineages Established Distribution}
\usage{
pmf.P_mu_est(n, R0, mu, maxMuGen, maxFS)
}
\arguments{
\item{n}{Initial population size, a positive integer}

\item{R0}{Reproduction number, a positive number}

\item{mu}{Mutation rate, a positive number}

\item{maxMuGen}{Maximum number of mutant lineages being calculated}

\item{maxFS}{Maximum number of final sizes being calculated}
}
\value{
A data frame of mutant lineages established and corresponding probabilities
}
\description{
This function calculates the mutant lineages established distribution.
}
\examples{
pmf.P_mu_est(2, 1.2, 0.4, 60, 60)
pmf.P_mu_est(3, 3.5, 1.6, 120, 130)
}
