% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClonalMuFunctions.R
\name{Px_analytical}
\alias{Px_analytical}
\title{Analytical Probability of Extinction}
\usage{
Px_analytical(n, R0)
}
\arguments{
\item{n}{Initial population size, a positive integer}

\item{R0}{Reproduction number, a positive number}
}
\value{
A number representing the probability
}
\description{
This function calculates the probability of the whole lineage goes extinction
under a geometric distribution assumption.
}
\examples{
Px_analytical(2, 1.2)
Px_analytical(3, 3.5)
}
